backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langgraph
langchain-openai
python-dotenv
requests
Pillow